feat: Improve navigation from homepage
- Move search box up
- Add standard dropdowns to header
